VCF loader split multi tests into separate file
